+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -31,7 +31,6 @@ type KurtosisBackend interface {
 		imageVersionTag string,
 		grpcPortNum uint16,
 		grpcProxyPortNum uint16,
-		engineDataDirpathOnHostMachine string,
 		envVars map[string]string,
 	) (
 		*engine.Engine,