+++ container-engine-lib/docs/changelog.md
@@ -4,6 +4,8 @@
     * Users no longer need to provide this argument
 * Removed `enclaveDataDirpathOnHostMachine` argument from `KurtosisBackend.CreateModule`
     * Users no longer need to provide this argument
+* Removed `engineDataDirpathOnHostMachine` from `KurtosisBackend.CreateEngine`
+    * Users no longer need to provide this argument
 
 # 0.18.0
 ### Features